requirements.txt
Flask==2.0.3
Flask-SocketIO==5.2.0
python-dotenv==0.19.2
openai==0.27.0
numpy==1.23.5
scikit-learn==1.2.0
sqlite3
